+++ src/App.css
@@ -115,6 +115,12 @@ button {
   width: 100%;
 }
 
+.header img {
+  width: 100%;
+  border-radius: 0;
+  min-width: 0;
+}
+
 /* Home Page */
 
 #home .header {
@@ -274,226 +280,4 @@ button {
 #instructions button {
   background-color: white;
   color: #0078ef;
-}
-
-/* Question Page */
-
-#question .header {
-  display: flex;
-  width: 100%;
-  align-items: flex-start;
-  justify-content: space-between;
-  padding-bottom: 20px;
-  border-bottom: 1px solid white;
-}
-
-.header img {
-  width: 100%;
-  border-radius: 0;
-  min-width: 0;
-}
-
-#question .question-box {
-  flex: 1 1 auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  gap: 20px;
-}
-
-.question-options {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 0px 10px;
-}
-
-.question-even {
-  gap: 10px;
-}
-
-#question button {
-  background-color: white;
-  color: #0078ef;
-  height: 92px;
-  width: 92px;
-  border-radius: 60px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 11px;
-}
-
-#question #get-location {
-  width: fit-content;
-  height: auto;
-  background-color: #0078ef;
-  color: white;
-  border: 1px solid white;
-  display: flex;
-  margin: 5px;
-  gap: 5px;
-}
-
-#question #get-location img {
-  height: 12px;
-  width: auto;
-}
-
-#question #logo-button {
-  padding: 0;
-  width: 40vw;
-  height: 40vw;
-  border-radius: 20vw;
-  margin-bottom: 10px;
-}
-
-#question .footer {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-}
-
-#question .footer button,
-#question #skip {
-  background-color: transparent;
-  color: white;
-  height: fit-content;
-  width: fit-content;
-  padding: 0px 0px 2px 0px;
-  border-radius: 0;
-  border-bottom: 1px solid white;
-  margin-bottom: 0px;
-}
-
-@media (min-width: 950px) {
-  #question .header {
-    padding-bottom: 30px;
-  }
-
-  #question .question-box {
-    gap: 30px;
-  }
-
-  .question-options {
-    gap: 0px 20px;
-  }
-
-  #question button {
-    height: 150px;
-    width: 150px;
-    border-radius: 100px;
-    font-size: 15px;
-  }
-
-  #question #logo-button {
-    width: 20vw;
-    height: 20vw;
-    border-radius: 10vw;
-  }
-}
-
-/* Final Page */
-
-#final .header {
-  width: 100%;
-}
-
-#final .food-links {
-  display: flex;
-  width: 100%;
-  gap: 8px;
-}
-
-#final .food-display {
-  flex: 1 1 auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 20px;
-  width: 100%;
-  min-height: 0;
-  position: relative;
-}
-
-#final .food-notes {
-  position: absolute;
-  top: 20px;
-  left: -10px;
-  background-color: #0078ef;
-  color: white;
-  padding: 10px 15px;
-  width: fit-content;
-  text-align: left;
-  border-radius: 5px;
-}
-
-#final #food-hearsay {
-  background-color: #ff4a31;
-}
-
-#final .meal-tags {
-  display: flex;
-  justify-content: center;
-  gap: 5px;
-  margin-top: 10px;
-}
-
-#final .meal-tags h4 {
-  color: white;
-  padding: 6px 10px;
-  width: fit-content;
-  height: 25px;
-  border-radius: 15px;
-}
-
-#final .meal-tags #veg {
-  background-color: #0fc771;
-}
-
-#final .meal-tags #halal {
-  background-color: #fd8156;
-}
-
-#final #logo-failed {
-  flex: 0 1 auto;
-  width: 50vw;
-  height: auto;
-}
-
-@media (min-width: 950px) {
-  #final #logo-failed {
-    width: 20vw;
-    min-width: 20vw;
-    height: 20vw;
-  }
-}
-
-#final .footer {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-@media (min-width: 950px) {
-  #final .food-display {
-    flex-direction: row;
-    gap: 20px;
-    width: 100%;
-    min-height: 0;
-  }
-
-  #final .food-display #food-details {
-    flex: 1;
-  }
-
-  #final .food-display img {
-    height: 100%;
-    min-width: 0;
-    flex: 3;
-  }
 }
