+++ src/Components/FinalScreen/FinalScreen.css
@@ -1,0 +1,100 @@
+#final .header {
+  width: 100%;
+}
+
+#final .food-links {
+  display: flex;
+  width: 100%;
+  gap: 8px;
+}
+
+#final .food-display {
+  flex: 1 1 auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 20px;
+  width: 100%;
+  min-height: 0;
+  position: relative;
+}
+
+#final .food-notes {
+  position: absolute;
+  top: 20px;
+  left: -10px;
+  background-color: #0078ef;
+  color: white;
+  padding: 10px 15px;
+  width: fit-content;
+  text-align: left;
+  border-radius: 5px;
+}
+
+#final #food-hearsay {
+  background-color: #ff4a31;
+}
+
+#final .meal-tags {
+  display: flex;
+  justify-content: center;
+  gap: 5px;
+  margin-top: 10px;
+}
+
+#final .meal-tags h4 {
+  color: white;
+  padding: 6px 10px;
+  width: fit-content;
+  height: 25px;
+  border-radius: 15px;
+}
+
+#final .meal-tags #veg {
+  background-color: #0fc771;
+}
+
+#final .meal-tags #halal {
+  background-color: #fd8156;
+}
+
+#final #logo-failed {
+  flex: 0 1 auto;
+  width: 50vw;
+  height: auto;
+}
+
+@media (min-width: 950px) {
+  #final #logo-failed {
+    width: 20vw;
+    min-width: 20vw;
+    height: 20vw;
+  }
+}
+
+#final .footer {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+@media (min-width: 950px) {
+  #final .food-display {
+    flex-direction: row;
+    gap: 20px;
+    width: 100%;
+    min-height: 0;
+  }
+
+  #final .food-display #food-details {
+    flex: 1;
+  }
+
+  #final .food-display img {
+    height: 100%;
+    min-width: 0;
+    flex: 3;
+  }
+}
